+++ src/UglyToad.PdfPig/Graphics/CurrentGraphicsState.cs
@@ -120,7 +120,7 @@ namespace UglyToad.PdfPig.Graphics
         /// <summary>
         /// The precision for rendering color gradients on the output device.
         /// </summary>
-        public decimal Smoothness { get; set; } = 0;
+        public double Smoothness { get; set; } = 0;
         #endregion
 
         /// <inheritdoc />
